backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets
sqlalchemy==2.0.25
alembic==1.13.1
python-binance==1.0.19
anthropic==0.18.1
google-genai==0.3.0
pandas==2.2.0
numpy==1.26.3
# ta-lib removed due to build issues
# pandas-ta removed due to install issues
feedparser==6.0.10
requests==2.31.0
beautifulsoup4==4.12.3
redis==5.0.1
hiredis==2.3.2
asyncio==3.4.3
aiohttp==3.9.1
cryptography==42.0.0
pyjwt==2.8.0
prometheus-client==0.19.0
python-dotenv==1.0.0
loguru==0.7.2
slowapi==0.1.9
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
pyyaml==6.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
schedule==1.2.1
python-dateutil==2.8.2
pytz==2024.1
scikit-learn==1.4.0
sentence-transformers==2.3.1
matplotlib==3.8.2
plotly==5.18.0
